mongodb web客户端 后台代码

+++ back/controllers/admin/schema.js
@@ -1,5 +1,5 @@
 const _ = require('lodash')
-const { Ctrl, ResultData } = require('tms-koa')
+const { Ctrl, ResultData, ResultFault } = require('tms-koa')
 const { Context } = require('../../context')
 const ObjectId = require('mongodb').ObjectId
 
@@ -67,10 +67,18 @@ class Schema extends Ctrl {
    */
   async remove() {
     const { id } = this.request.query
-    const client = await Context.mongoClient()
-    return client
-      .db('tms_admin')
-      .collection('mongodb_object')
+    if (!id) return new ResultFault("参数不完整")
+
+    const client = this.mongoClient
+    const cl = client.db('tms_admin').collection('mongodb_object')
+    
+    // 是否正在使用
+    let rst = await cl.findOne({schema_id: id, type: "collection"})
+    if (rst) {
+      return new ResultFault("集合列正在被使用不能删除")
+    }
+
+    return cl
       .deleteOne({ _id: new ObjectId(id), type: 'schema' })
       .then(() => new ResultData('ok'))
   }
